rework the color code some more
